Account Creation Now possible
I am still figuring out what to do about API keys. For firebase, mine are stored in a firebaseConfig.js file in the same directory as App.js.

In this commit I:
Created a login page
Created a create account page
Made it so the home page now displays the user's name