Profile Picture Changeable, Assets in Database
The application is now 100% functional without the metro server running!

Thanks for Storage in Firebase, the logo.png and avatar.png are now stored on the cloud!

I have also made it so that you can change the profile picture.

This was done using ImagePicker from Expo. When the iamge is selected,
it is uploaded to Firebase Storage using uploadByteResumable. The app then stores the image's path in Firestore.